// src/app/shared/interfaces/auxilium/company-center/company.interface.ts
export interface CompanyDisplay {
    id: number;
    type: string;
    name: string;
    address: string;
    city: string;
    state: string;
    zip: string;
    fedid: string;
    phone: string;
    fax: string;
    lastpat: number;
    lastorder: number;
    lastrx: number;
    mcaresupl: string;
    mcaresend: string;
    mcarermt: string;
    mcaidid: string;
    kct: string;
    specialpfx: string;
    dmercid: string;
    remote: string;
    sender: string;
    rate: string;
    participat: string;
    agevar: string;
    ldate: string;
    cmd1: string;
    cmd2: string;
    cmd3: string;
    cmd4: string;
    dmerc: string;
    batchno: number;
    contactfirstname: string;
    contactlastname: string;
    exportpath: string;
    importpath: string;
    invoicemessage: string;
    medicaidexportpath: string;
    ha0default: string;
    autoadd: boolean;
    form1500: string;
    backupdriveandpath: string;
    duns: string;
    taxonomy: string;
    ack997: string;
    tax: number;
    notes: string;
    multibranch: boolean;
    startupform: string;
    scanner: boolean;
    alerts: boolean;
    patspecial: boolean;
    taxcounty: number;
    taxlocal: number;
    billrefer: boolean;
    defaultadjustmentcode: string;
    allicd9codes: boolean;
    broadcast: string;
    rxinitials: string;
    asktosave: boolean;
    defaultbilltoid: number;
    submittername: string;
    submitteremail: string;
    diabcoaccount: string;
    homedir: string;
    npi: string;
    npionly: boolean;
    defaultadjcode: string;
    email: string;
    website: string;
    sqlserver: string;
    epoexportpath: string;
    epoimportpath: string;
    vgexportpath: string;
    vgimportpath: string;
    realtimeedi: string;
    ediuser: string;
    edipw: string;
    ssrs: string;
}
export type GetCompaniesResponse = CompanyDisplay[];